requirements.txt
numpy==1.15.4
Pillow==5.3.0
pyopencl==2018.1.1
